Patch a performance issue with hard puzzles that require multiple branchings
In the shallow copy of the working grid done when branching, copy the
nb_alternatives vectors instead of resetting it.

Failing to do so means that on subsequent branchings, we might miss the
opportunity to branch on a row/column with a small number of
alternatives. As a consequence the total processing time hugely
increases by the combinatorial effect of having several branching points
with a high number of alternatives to test.