Refactor LineAlternatives and unit test it
The line constraints can now be processed in one direction or the other.
For instance, for a row constraint, from left to right or from right to
left.

This is a step in preparation for a future change where only a subpart of
the line constraint (e.g. the leftmost part or the rightmost part) is
used to reduce the line.