Use memoization in full_reduction line solver
The gain in performance is worth the cost in memory. For memoization
we need a O(k.n^2) buffer where k is the number of constraints on the
line and n is the line size. The full reduction is now in O(k.n^3)
time complexity and there is still room for improvement.