No limit on max_nb_alternatives at search depth 0
Now that the speed of the full_reduction as been greatly improved,
there is no reason to limit the max_nb_alternatives on the "ref" solver
(by opposition with the "line" solver).

As a consequence, line-solvable grids will always be identified as such.
The difference in behavior is that the "line" solver will not do a
branch search. It will stop at depth 0 and return the partial grid
corresponding to what could be solved using a line solver only.