Remove filled_tiles_masks from linear_reduction
Those masks allow to find more filled tiles in the linear_reduction, but
at the cost at some cpu time and memory (it requires a O(n^2) size
buffer).

Given the significant progress on the full_reduction, that does not seem
worth it anymore. The goal of the linear reduction is to make fast
progress in the resolution of the puzzle.

This is confirmed by the benchmark, especially on the puzzles from
nonograms.org, which can be quite complex but still line solvable.
E.g. "Gargantua".