CLI: Add the --output option
Issue #1

With that option, the CLI will save a grid file, in a format that is
deduced from the output file extension. If a solution is found, and
the output format supports it, then that solution is included in the
saved file.

Caveats:
- If there are multiple grids in the input, the output file
gets overwritten. For that reason the --output option should only be
used when solving a single input.
- If the input grid has multiple solutions, the first one being found
will be used as the goal in the saved file.